Remove the saddle -- and got it to compile

+++ app/analysis/junkyard/SpreadDist.scala
@@ -1,4 +1,4 @@
 package analysis.junkyard
 
 
-case class SpreadDist(name:String, n:Int, mean:Double, stDev:Double, skew:Double, kurtosis:Double, pctOnePtGames:Double, min:Int, q1:Double, med:Double, q3:Double, p95:Double, p99:Double, max:Int, mode:(List[Int], Int))
+case class SpreadDist(name:String, n:Double, mean:Double, stDev:Double, skew:Double, kurtosis:Double, pctOnePtGames:Double, min:Double, q1:Double, med:Double, q3:Double, p95:Double, p99:Double, max:Double, mode:(List[Int], Int))
